+++ client/excel-addin/deltic-excel-addin/src/commands/commands.js
@@ -1,0 +1,10 @@
+Office.onReady(() => {
+  console.log("Commands runtime ready");
+});
+
+/**
+ * Called by the ribbon button via ExecuteFunction
+ */
+function showTaskpane() {
+  Office.addin.showAsTaskpane();
+}